requirements.txt
streamlit>=1.36
pandas>=2.0
altair>=5.1.2
vl-convert-python>=1.5.0
